splitting up files and renaming stuff
